delete orders, product

+++ orders/index.php
@@ -1,23 +1,25 @@
 <?php include("../server/conn.php");
-    
-        if(isset($_POST['btn-search'])){
-            $search = $_POST['txt-search'];
-            $query = "SELECT * FROM orders WHERE order_id LIKE '%$search%'";
-        } else {
-            $query = "SELECT * FROM orders";
-        }
-    
-        $result = mysqli_query($conn, $query);
+
+if (isset($_POST['btn-search'])) {
+    $search = $_POST['txt-search'];
+    $query = "SELECT * FROM orders WHERE order_id LIKE '%$search%'";
+} else {
+    $query = "SELECT * FROM orders";
+}
+
+$result = mysqli_query($conn, $query);
 ?>
 
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Manage Orders</title>
 </head>
+
 <body>
     <div class="container">
         <div class="top">
@@ -40,11 +42,11 @@
                         <th>Order Quantity</th>
                         <th>Total Price</th>
                         <th>Date</th>
-                        <th>Action</th> 
+                        <th>Action</th>
                     </tr>
                 </thead>
                 <tbody>
-                    <?php while($row = mysqli_fetch_array($result)): ?>
+                    <?php while ($row = mysqli_fetch_array($result)) : ?>
                         <tr>
                             <td><?= $row['order_id'] ?></td>
                             <td><?= $row['admin_id'] ?></td>
@@ -53,8 +55,12 @@
                             <td><?= $row['order_cost'] ?></td>
                             <td><?= $row['date'] ?></td>
                             <td>
-                                <input type="submit" class="btn-update" value="Edit" onclick="location.href='update.php?order_id=<?= $row['order_id'] ?>';">
-                                <input type="submit" class="btn-delete" value="delete" onclick="location.href='delete.php?order_id=<?= $row['order_id'] ?>';">
+                                <div class="alert">
+                                    <input type="submit" class="btn-update" value="Edit" onclick="location.href='update.php?order_id=<?= $row['order_id'] ?>';">
+                                    <span class="closebtn" onclick="return confirm ('Data ini akan di hapus?')">
+                                        <input type="submit" class="btn-delete" value="delete" onclick="location.href='delete.php?order_id=<?= $row['order_id'] ?>';">
+                                    </span>
+                                </div>
                             </td>
                         </tr>
                     <?php endwhile; ?>
@@ -64,4 +70,5 @@
 
     </div>
 </body>
+
 </html>
